Add application for reading data from serial port
./run.sh will create a Docker image and spin up a container with the application.

The application itself:
- Creates 2 virtual serial ports using satcom.
- Starts a bash wrapper script that writes "dummy" entries (every 10 seconds) to the first serial port,
which then passes the data to the second one.
- Starts a python script that serves as an entry point for the application.
This one will be monitoring what's written to the second serial port and then
write the entries to a log file.

Notes:
- The source (the second serial virtual port) could be changed using the SOURCE environment
variable inside run.sh. The source could also be a physical serial port, then satcom's functionality
will not take effect and the data will be entirely monitored from there, instead of the virtual port.
- The file (destination file the log messages coming from the source are written to) could also be changed
using the LOG_FILE environment variable inside run.sh.
- The LOG_FILE file is also mounted at the current working directory, so it could be easily accessed
outside the container - ./device-logs/$LOG_FILE.

ToDo:
Write a gitlab-ci.yaml to include the image in a CI job.