Populate .gitlab-ci.yml
.gitlab-ci.yml could now be used to run a pipeline in GitLab which will deploy the docker container in a runner in GitLab.
The image was created and pushed in advance to my personal Docker repisitory in Dockerhub.
Also, due to the nature of how the python script was created - to run until manually terminated - I've updated the application to simulate
only 10 messages to the log entries and then terminate on its own. This will make sure an artifact could be collected in GitLab, in conjuction with the
artifact/when: always parameter as specified in .gitlab-ci.yml